Rename projects to better reflect the book
