+++ csharp2/Dockerfile
@@ -1,22 +1,21 @@
 ﻿FROM mcr.microsoft.com/dotnet/sdk:7.0-alpine AS base
 WORKDIR /app
-EXPOSE 80
 
 FROM mcr.microsoft.com/dotnet/sdk:7.0-alpine AS restore
 WORKDIR /src
-COPY ["lexer/lexer.csproj", "lexer/"]
-COPY ["lexer.tests/lexer.tests.csproj", "lexer.tests/"]
-RUN dotnet restore --ucr "lexer/lexer.csproj" && \
-    dotnet restore --ucr "lexer.tests/lexer.tests.csproj"
+COPY ["Monkey/Monkey.csproj", "Monkey/"]
+COPY ["MonkeyTests/MonkeyTests.csproj", "MonkeyTests/"]
+RUN dotnet restore --ucr "Monkey/Monkey.csproj" && \
+    dotnet restore --ucr "MonkeyTests/MonkeyTests.csproj"
 
 FROM restore AS publish
 COPY . .
-RUN dotnet publish -c release --no-restore "lexer/lexer.csproj" \
+RUN dotnet publish -c release --no-restore "Monkey/Monkey.csproj" \
     -p:publishtrimmed=true --ucr -o /app/publish && \
-    dotnet publish -c release --no-restore "lexer.tests/lexer.tests.csproj" \
+    dotnet publish -c release --no-restore "MonkeyTests/MonkeyTests.csproj" \
     --ucr -o /app/tests
 
 FROM base as final
-COPY --from=publish /app/publish ./lexer
+COPY --from=publish /app/publish ./app
 COPY --from=publish /app/tests ./tests
-ENTRYPOINT ["./lexer/lexer"]
+ENTRYPOINT ["./app/Monkey"]
