+++ csharp2/Makefile
@@ -1,23 +1,23 @@
-IMAGE=lexer-csharp2
+IMAGE=monkey-csharp2
 
 all:
 	@echo "Commands: bench, clean, fmt, test, ready, docker-build, docker-ready"
 
 bench:
-	@dotnet run -c release --project lexer.bench
+	@dotnet run -c release --project MonkeyBench
 
 clean:
-	@dotnet clean lexer
-	@dotnet clean lexer.tests
-	@dotnet clean lexer.bench
+	@dotnet clean Monkey
+	@dotnet clean MonkeyTests
+	@dotnet clean MonkeyBench
 
 fmt:
-	@dotnet format lexer
-	@dotnet format lexer.tests
-	@dotnet format lexer.bench
+	@dotnet format Monkey
+	@dotnet format MonkeyTests
+	@dotnet format MonkeyBench
 
 test:
-	@dotnet test lexer.tests
+	@dotnet test MonkeyTests
 
 ready: fmt test
 
@@ -25,4 +25,4 @@ docker-build:
 	@docker build -t ${IMAGE} .
 
 docker-ready: docker-build
-	@docker run --rm -ti --entrypoint "dotnet" ${IMAGE} test tests/lexer.tests.dll
+	@docker run --rm -ti --entrypoint "dotnet" ${IMAGE} test tests/MonkeyTests.dll
