+++ csharp2/MonkeyBench/MonkeyBench.csproj
@@ -11,6 +11,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\lexer\lexer.csproj" />
+    <ProjectReference Include="..\Monkey\Monkey.csproj" />
   </ItemGroup>
 </Project>
