+++ csharp2/MonkeyTests/MonkeyTests.csproj
@@ -22,6 +22,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\lexer\lexer.csproj" />
+    <ProjectReference Include="..\Monkey\Monkey.csproj" />
   </ItemGroup>
 </Project>
